fix delete

+++ app/Repositories/Api/User/EloquentGuideTripUserApiRepository.php
@@ -104,9 +104,16 @@ class EloquentGuideTripUserApiRepository implements GuideTripUserApiRepositoryIn
     public function deleteSubscriberInTrip($slug)
     {
         $guideTrip = GuideTrip::findBySlug($slug);
-        $guideTripUser = GuideTripUser::where('guide_trip_id', $guideTrip->id)->where('user_id', Auth::guard('api')->user()->id)->get();
-        $guideTripUser->delete();
+
+        $guideTripUser = GuideTripUser::where('guide_trip_id', $guideTrip->id)
+            ->where('user_id', Auth::guard('api')->user()->id)
+            ->first();
+
+        if ($guideTripUser) {
+            $guideTripUser->delete();
+        }
     }
+
 
     public function allSubscription($slug)
     {
